Refactor font sizes to store pixel values directly
Eliminates unnecessary conversion roundtrip:
- Before: CSS 48px → ÷8 → level 6 → lookup table → 48px
- After: CSS 48px → store 48.0 → use 48.0

Changes:
- LayoutBox.font_size: usize → f32 (stores actual pixels)
- Deleted get_font_size_px() lookup table (23 lines)
- Layout stores pixel values directly (h1=36.0, h2=28.0, etc.)
- Bitmap font calculates multiplier on-the-fly: (px/8.0 + 0.5) as usize
- TTF font uses pixel values directly

Benefits:
- No magic number 8 scattered everywhere
- Self-documenting: font_size=48.0 vs font_size_level=6
- Single source of truth (CSS pixel values)
- Cleaner code per CLAUDE.md standards

+++ src/gui/widgets/browser/types.rs
@@ -117,7 +117,7 @@ pub struct LayoutBox {
     pub text: String,
     pub color: Color,
     pub background_color: Option<Color>, // CSS background-color
-    pub font_size: usize, // Multiplier for 8x8 font (1=8px, 2=16px, etc.)
+    pub font_size: f32, // Font size in pixels (e.g., 18.0, 48.0)
     pub is_link: bool,
     pub link_url: String,
     pub bold: bool,
